fix(engine): resolve critical search, UCI, and board safety bugs (+66.8 Elo)
- UCI & Threading: Offloaded searcher->think() to a background thread to prevent blocking UCI input commands (stop/quit).
- Transposition Table: Upgraded TT entry depth from 8-bit to 16-bit to prevent integer overflow/wrap at deep plies.
- Search Safety: Replaced unsafe fallback moves with isLegalMove() and pickLegalMove() validation to guarantee zero illegal moves on checkmate/stalemate.
- Bitboard Protection: Guarded bsf()/getlsb() calls against empty king bitboards (0ULL) to prevent hardware instruction undefined behavior.
- FEN Parsing: Added file/rank boundary validation to prevent buffer overflows from malformed FEN strings.
- Castling MoveGen: Explicitly verified King and Rook presence on starting squares before generating castling moves.
- Search Abort Handling: Prevented saving aborted search nodes to TT and fixed early exit logic in Iterative Deepening.
- Limits Enforcement: Added node count limit check to CheckTime() for full UCI  support.

Benchmark vs ASTROVE 1.0 (UHO_2024 8-move EPD, 200 games @ 8+0.08):
- Score: 62W – 24L – 114D (59.5% win rate)
- Elo Difference: +66.8 ± 31.4

+++ src/table/tt.cpp
@@ -90,7 +90,7 @@ void TranspositionTable::store(uint64_t key, int depth, int flag, int score,
   }
 
   replace->key = key;
-  replace->depth = static_cast<int8_t>(depth);
+  replace->depth = static_cast<int16_t>(depth);
   replace->flag = static_cast<uint8_t>(flag);
   replace->score = static_cast<int16_t>(score);
   replace->eval = static_cast<int16_t>(eval);
